+++ src/common/TodoBase.ts
@@ -1,6 +1,4 @@
 export default abstract class TodoBase {
   public update(){}
-  public configChanged(){
-    console.log('c')
-  }
+  public configChanged(){}
 }